refactor(training): simplify MoE auxiliary loss scaling
