+++ nemo_automodel/components/training/utils.py
@@ -375,76 +375,6 @@ def get_expert_tp_replication_factor(
     return 1
 
 
-def get_moe_aux_loss_backward_scale(
-    *,
-    num_microbatches: int,
-    cp_group_size: int,
-    pp_enabled: bool,
-    num_label_tokens: int | None = None,
-    dp_group_size: int | None = None,
-) -> float:
-    """Return the per-microbatch MoE auxiliary-loss backward scale.
-
-    The load-balancing loss is already averaged over the tokens in one model
-    microbatch. Its gradients therefore need to be averaged over all model
-    microbatches in an optimizer step. FSDP averages gradients over the
-    flattened DP-CP group, so the scale also restores the CP sum while keeping
-    the DP average.
-
-    Pipeline-parallel recipes normalize all gradients after backward by
-    ``num_label_tokens / dp_group_size``. The returned PP scale compensates for
-    that post-backward normalization exactly once. When a PP step has no label
-    tokens, the post-backward normalization is skipped and the base scale is
-    returned unchanged.
-
-    Args:
-        num_microbatches: Number of model forward/backward microbatches in the
-            current optimizer step, including PP schedule microbatches.
-        cp_group_size: Number of context-parallel ranks contributing token
-            shards to each auxiliary loss.
-        pp_enabled: Whether pipeline parallelism and its post-backward token
-            normalization are enabled.
-        num_label_tokens: Global number of supervised tokens in the optimizer
-            step. Required when ``pp_enabled`` is true.
-        dp_group_size: Size of the flattened DP-CP gradient-reduction group.
-            Required when ``pp_enabled`` is true.
-
-    Returns:
-        Scale applied to each microbatch's auxiliary-loss gradient.
-
-    Raises:
-        TypeError: If a size or token count is not an integer.
-        ValueError: If a size is invalid or PP normalization inputs are absent.
-    """
-
-    for name, value in (("num_microbatches", num_microbatches), ("cp_group_size", cp_group_size)):
-        if not isinstance(value, int) or isinstance(value, bool):
-            raise TypeError(f"{name} must be an integer")
-        if value < 1:
-            raise ValueError(f"{name} must be >= 1")
-
-    scale = float(cp_group_size) / float(num_microbatches)
-    if not pp_enabled:
-        return scale
-
-    if num_label_tokens is None or dp_group_size is None:
-        raise ValueError("num_label_tokens and dp_group_size are required when pp_enabled is True")
-    for name, value, minimum in (
-        ("num_label_tokens", num_label_tokens, 0),
-        ("dp_group_size", dp_group_size, 1),
-    ):
-        if not isinstance(value, int) or isinstance(value, bool):
-            raise TypeError(f"{name} must be an integer")
-        if value < minimum:
-            raise ValueError(f"{name} must be >= {minimum}")
-    if dp_group_size % cp_group_size != 0:
-        raise ValueError("dp_group_size must include and be divisible by cp_group_size")
-
-    if num_label_tokens > 0:
-        scale *= float(num_label_tokens) / float(dp_group_size)
-    return scale
-
-
 @torch.no_grad()
 def scale_grads_and_clip_grad_norm(
     max_grad_norm: float | None,
