+++ nemo_automodel/recipes/base_recipe.py
@@ -63,7 +63,6 @@ from nemo_automodel.components.optim.scheduler import OptimizerParamScheduler
 from nemo_automodel.components.training.garbage_collection import GarbageCollection
 from nemo_automodel.components.training.rng import StatefulRNG
 from nemo_automodel.components.training.step_scheduler import StepScheduler
-from nemo_automodel.components.training.utils import get_moe_aux_loss_backward_scale
 from nemo_automodel.recipes._typed_config import RecipeConfig
 
 logger = logging.getLogger(__name__)
@@ -964,26 +963,20 @@ class BaseRecipe:
         return device_mesh["cp"].size()
 
     def _set_moe_aux_loss_backward_scale(self, *, num_batches: int, num_label_tokens: int) -> None:
-        """Set the MoE auxiliary-loss scale for one optimizer step.
+        """Set the per-microbatch MoE auxiliary-loss scale for one optimizer step.
 
-        Args:
-            num_batches: Number of outer gradient-accumulation batches in the
-                optimizer step.
-            num_label_tokens: Global number of supervised tokens in the
-                optimizer step.
+        The base scale averages accumulation microbatches and restores the CP
+        sum lost in the flattened DP-CP gradient average. PP additionally needs
+        to compensate for its post-backward token normalization.
         """
         num_model_microbatches = num_batches
         if self.pp_enabled:
-            num_model_microbatches *= self.pp.pp_batch_size // self.pp.pp_microbatch_size
-        MoEAuxLossAutoScaler.main_loss_backward_scale = torch.tensor(
-            get_moe_aux_loss_backward_scale(
-                num_microbatches=num_model_microbatches,
-                cp_group_size=self._get_cp_group_size(),
-                pp_enabled=self.pp_enabled,
-                num_label_tokens=num_label_tokens,
-                dp_group_size=self._get_dp_group_size(include_cp=True),
-            )
-        )
+            num_model_microbatches *= self.pp.info.schedule.n_microbatches
+
+        scale = self._get_cp_group_size() / num_model_microbatches
+        if self.pp_enabled and num_label_tokens > 0:
+            scale *= num_label_tokens / self._get_dp_group_size(include_cp=True)
+        MoEAuxLossAutoScaler.main_loss_backward_scale = torch.tensor(scale)
 
     def _get_dp_rank(self, include_cp: bool = False):
         device_mesh = getattr(self, "device_mesh", None)
